Process Monitor - simplified structure, moved everything to Core
