+++ src/Artemis.UI/Bootstrapper.cs
@@ -89,7 +89,6 @@ namespace Artemis.UI
 
             IRegistrationService registrationService = Kernel.Get<IRegistrationService>();
             registrationService.RegisterInputProvider();
-            registrationService.RegisterProcessWatchers();
             registrationService.RegisterControllers();
             
             // Initialize background services
